+++ packages/yourloops/components/theme.tsx
@@ -64,6 +64,13 @@ export function getTheme(): Theme {
         }
       `
       },
+      MuiAlert: {
+        styleOverrides: {
+          root: {
+            borderRadius: 12
+          }
+        }
+      },
       MuiButton: {
         styleOverrides: {
           root: {