+++ packages/yourloops/layout/hcp-layout-with-context.tsx
@@ -26,29 +26,16 @@
  */
 
 import React, { type FunctionComponent } from 'react'
-import { SimpleStat, StatFormats } from 'dumb'
-import { t } from 'i18next'
-import Box from '@mui/material/Box'
-import { type BgType, DatumType } from 'medical-domain'
+import { SelectedTeamProvider } from '../lib/selected-team/selected-team.provider'
+import { HcpLayout } from './hcp-layout'
+import { TeamContextProvider } from '../lib/team'
 
-export interface CoefficientOfVariationProps {
-  coefficientOfVariation: number
-  bgType: BgType
-}
-
-export const CoefficientOfVariation: FunctionComponent<CoefficientOfVariationProps> = (props) => {
-  const { coefficientOfVariation, bgType } = props
-  const annotation = coefficientOfVariation ? [t('coefficient-of-variation-tooltip')] : [t('coefficient-of-variation-tooltip'), t('tooltip-empty-stat')]
-  const selectedLabel = bgType === DatumType.Cbg ? t('CGM') : t('BGM')
+export const HcpLayoutWithContext: FunctionComponent = () => {
   return (
-    <Box data-testid="coefficient-of-variation-stat">
-      <SimpleStat
-        annotations={annotation}
-        title={t('coefficient-of-variation', { cbgLabel: selectedLabel })}
-        value={coefficientOfVariation}
-        summaryFormat={StatFormats.Cv}
-        total={0}
-      />
-    </Box>
+    <TeamContextProvider>
+      <SelectedTeamProvider>
+        <HcpLayout />
+      </SelectedTeamProvider>
+    </TeamContextProvider>
   )
 }